In progress work

+++ com.unity.render-pipelines.universal/Runtime/2D/Shadows/IShadowShapes2DProvider.cs
@@ -17,7 +17,7 @@ namespace UnityEngine.U2D
         public abstract class ShadowShapes2D
         {
             public abstract void SetEdges(Vector2[] vertices, ushort[] indices, OutlineTopology outlineTopology);
-            public abstract void SetEdges(NativeArray<Vector2> vertices, NativeArray<ushort> indices, OutlineTopology outlineTopology);
+            public abstract void SetEdges(NativeArray<Vector2> vertices, NativeArray<int> indices, OutlineTopology outlineTopology);
             public abstract void UpdateEdges(Vector2[] vertices);
             public abstract void UpdateEdges(NativeArray<Vector2> vertices);
         }
